requirements.txt
aioredis==2.0.1
antlr4-python3-runtime==4.9.2
anyio==3.6.2
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
async-timeout==4.0.2
awscli==1.22.34
bcrypt==3.2.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dask==2023.4.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-limiter==0.1.5
fasteners==0.14.1
filelock==3.9.0
fire==0.5.0
fsspec==2023.4.0
future==0.18.2
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httplib2==0.20.2
httpx==0.24.0
idna==3.3
importlib-metadata==6.6.0
iniconfig==2.0.0
jeepney==0.7.1
jmespath==0.10.0
keyring==23.5.0
language-selector==0.1
lark-parser==0.12.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.39.1
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
packaging==23.1
PAM==0.4.2
paramiko==2.9.3
partd==1.4.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.6.2
pluggy==1.0.0
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.7
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.3.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
quantum-blackbird==0.5.0
quantum-xir==0.2.2
redis==4.5.4
reportlab==3.6.8
requests==2.25.1
requests-unixsocket==0.2.0
roman==3.3
rsa==4.8
s3transfer==0.5.0
scipy==1.10.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
strawberry-graphql==0.171.3
sympy==1.11.1
systemd-python==234
termcolor==2.2.0
thewalrus==0.20.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.21.1
virtualenv==20.17.1
wadllib==1.3.6
xanadu-cloud-client==0.3.0
xdg==5
xkit==0.0.0
zipp==1.0.0
